Fix various linux/mac conflicts

+++ bash/functions.bash
@@ -1,13 +1,13 @@
 # Generic Bash Functions
 
 ## Enable autojump: https://github.com/joelthelion/autojump
-if [ -f `brew --prefix`/etc/autojump ]; then
-  . `brew --prefix`/etc/autojump
+if [ -f $BREWDIR/etc/autojump ]; then
+  . $BREWDIR/etc/autojump
 fi
 
 ## History Search Shorthand
 function hist {
-    history | grep $1
+  history | grep $1
 }
 
 ## CD and immediately ls
@@ -20,10 +20,14 @@ function cdls {
 alias sshadd='ssh-add ~/.ssh/id_rsa'
 
 ## Guess what I meant
-alias f='$(thefuck $(fc -ln -1))'
+if [[ -f `which thefuck` ]]; then
+  alias f='$(thefuck $(fc -ln -1))'
+fi
 
 # VirtualBox management
-alias vb=VBoxManage
+if [[ -f `which VBoxManage` ]]; then
+  alias vb=VBoxManage
+fi
 
 # Handy Python web-servers: Usage `serve [port]`
 function http {
@@ -43,20 +47,25 @@ function killport {
 alias urlencode='python -c "import sys, urllib as ul; print ul.quote_plus(sys.argv[1])"'
 
 # Open Sublime Project
-function subl {
-  if [ -n "$1" ]; then
-    command subl $1
-  elif [ -f *.sublime-project ]; then
-    command subl *.sublime-project
-  else
-    command subl .
-  fi
-}
+if [[ -f `which subl` ]]; then
+  function subl {
+    if [ -n "$1" ]; then
+      command subl $1
+    elif [ -f *.sublime-project ]; then
+      command subl *.sublime-project
+    else
+      command subl .
+    fi
+  }
+fi
+
 # Open vscode directory
-function code {
-  if [ -n "$1" ]; then
-    command code $1
-  else
-    command code .
-  fi
-}
+if [[ -f `which code` ]]; then
+  function code {
+    if [ -n "$1" ]; then
+      command code $1
+    else
+      command code .
+    fi
+  }
+fi
